Add a large realistic fixture to test ngs with
